# api-gateway/docs/swagger.yaml
definitions:
...
  pb.AbsGetHistoryByUserIdResponse:
    properties:
      createdAt:
        type: string
      currentValue:
        type: string
      oldValue:
        type: string
      updatedField:
        type: string
    type: object
...
  pb.AbsTeacher:
    properties:
      activeGroups:
        type: string
      fullName:
        type: string
      id:
        type: string
      phoneNumber:
        type: string
    type: object
...
  pb.CreateUserRequest:
    properties:
      birthDate:
        type: string
      fullName:
        type: string
      gender:
        type: boolean
      password:
        type: string
      phoneNumber:
        type: string
      role:
        type: string
    type: object
...
  pb.GetAllEmployeeResponse:
    properties:
      employees:
        items:
          $ref: '#/definitions/pb.GetUserByIdResponse'
        type: array
    type: object
...
  pb.GetAllStuffResponse:
    properties:
      stuff:
        items:
          $ref: '#/definitions/pb.GetUserByIdResponse'
        type: array
    type: object
...
  pb.GetHistoryByUserIdResponse:
    properties:
      histories:
        items:
          $ref: '#/definitions/pb.AbsGetHistoryByUserIdResponse'
        type: array
    type: object
...
  pb.GetTeachersResponse:
    properties:
      teachers:
        items:
          $ref: '#/definitions/pb.AbsTeacher'
        type: array
    type: object
...
  pb.GetUserByIdResponse:
    properties:
      birthDate:
        type: string
      companyId:
        type: integer
      createdAt:
        type: string
      gender:
        type: boolean
      id:
        type: string
      is_deleted:
        type: boolean
      name:
        type: string
      phoneNumber:
        type: string
      role:
        type: string
    type: object
...
  pb.LoginRequest:
    properties:
      password:
        type: string
      phoneNumber:
        type: string
    type: object
  pb.LoginResponse:
    properties:
      isOk:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/pb.GetUserByIdResponse'
    type: object
...
  pb.UpdateUserRequest:
    properties:
      birthDate:
        type: string
      gender:
        type: boolean
      id:
        type: string
      name:
        type: string
      phoneNumber:
        type: string
      role:
        type: string
    type: object
  utils.AbsResponse:
    properties:
      message:
        type: string
      statusCode:
        type: integer
    type: object
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Sphere Swagger
paths:
...
  /api/user/create:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/pb.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created user
          schema:
            $ref: '#/definitions/utils.AbsResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.AbsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      security:
      - Bearer: []
      summary: CEO
      tags:
      - user
  /api/user/delete/{userId}:
    delete:
      description: Delete a user from the system using their ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.AbsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      security:
      - Bearer: []
      summary: ADMIN , CEO
      tags:
      - user
  /api/user/get-all-employee/{isArchived}:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all employees based on archive status. Restricted
        to ADMIN and CEO roles.
      parameters:
      - description: Filter by archive status (true=archived, false=active)
        in: path
        name: isArchived
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.GetAllEmployeeResponse'
        "400":
          description: Invalid input or processing error
          schema:
            $ref: '#/definitions/utils.AbsResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/utils.AbsResponse'
        "408":
          description: Request timeout
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      security:
      - Bearer: []
      summary: ADMIN , CEO
      tags:
      - user
  /api/user/get-all-staff/{isArchived}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all staff members, filtered by their archived
        status.
      parameters:
      - description: Boolean to filter archived or active staff; true for archived,
          false for active
        in: path
        name: isArchived
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of staff members based on archived status
          schema:
            items:
              $ref: '#/definitions/pb.GetAllStuffResponse'
            type: array
        "400":
          description: Bad request - Invalid 'isArchived' parameter
          schema:
            $ref: '#/definitions/utils.AbsResponse'
        "500":
          description: Internal server error during data retrieval
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      security:
      - Bearer: []
      summary: ADMIN , CEO
      tags:
      - user
  /api/user/get-my-profile:
    get:
      consumes:
      - application/json
      description: Retrieve the information of the authenticated user. Admin and CEO
        roles receive basic user info, while Teachers get additional group information.
      produces:
      - application/json
      responses:
        "200":
          description: User information with optional groups for Teachers
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or missing authentication
          schema:
            $ref: '#/definitions/utils.AbsResponse'
        "500":
          description: Internal server error while retrieving teacher's group information
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      security:
      - Bearer: []
      summary: ADMIN , CEO , TEACHER
      tags:
      - user
  /api/user/get-teachers/{isDeleted}:
    get:
      consumes:
      - application/json
      description: Fetches a list of teachers based on the deletion status
      parameters:
      - description: Deletion status (true/false)
        in: path
        name: isDeleted
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of teachers
          schema:
            $ref: '#/definitions/pb.GetTeachersResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.AbsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      security:
      - Bearer: []
      summary: ADMIN,CEO
      tags:
      - user
  /api/user/get-user/{userId}:
    get:
      description: Retrieve user details by their unique user ID.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.GetUserByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      summary: CEO , ADMIN , TEACHER
      tags:
      - user
  /api/user/history/{userId}:
    get:
      description: Fetches the history data for a given user ID.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.GetHistoryByUserIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      security:
      - Bearer: []
      summary: ADMIN, CEO, TEACHER
      tags:
      - user
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a token upon successful login.
      parameters:
      - description: Login credentials
        in: body
        name: LoginRequest
        required: true
        schema:
          $ref: '#/definitions/pb.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            $ref: '#/definitions/pb.LoginResponse'
        "400":
          description: Bad request - Invalid JSON or login failure
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      summary: ALL
      tags:
      - user
  /api/user/update:
    patch:
      consumes:
      - application/json
      description: Update user details using their ID
      parameters:
      - description: User Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/pb.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.AbsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      summary: ADMIN , CEO , TEACHER
      tags:
      - user
  /api/user/update-password/{userId}/{password}:
    put:
      consumes:
      - application/json
      description: Updates the password of a user specified by the userId.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: New Password
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            $ref: '#/definitions/utils.AbsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AbsResponse'
      security:
      - Bearer: []
      summary: CEO
      tags:
      - user
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"